Overwrite calculated values in valuesHash
If the name of a calculated value with no dependents, but where
includeIntoResult = true, is present in the valuesHash (because it was
loaded from saved data), the old value is returned from survey.data.
This change moves the writing of calculated values to after copying
from valuesHash, so that the new result is reflected.